Directory Restructuring
Moved all of the js functoins into a file, and all of the non program specific functions into another file. Declared both of those and all of the SVG css in an SVG container in the head. Leaving only event handler additions in the head and body SVG divs. Also fixed the power titles to expand and shift a bit on the hovering of their box.