Year labels done
Year labels on the y axis of the chart are now moving and sticking perfectly. I also toyed around with moving all of the variable declarations into a separate javascript file, but I couldn't get things to work correctly with how things need to be defined in SVG spaces. It was weird. But I left it should I want to try again later.