Header bar movement adjusted
The header bar no longer gets taller than a scale of 1, the texts are counter scaled to not stretch, and then translated to the center of their boxes. The truncation of the region text also dynamically changes for the changing width of the box. Finally, the texts are not truncated in python, and are run through the length adjustment on page load. This was causing a flickering during load where the full length is loaded and then changed. So The header texts start invisible and are toggled to visible on initial load.