Power Focus Feature Done
If you double click on a power it makes all other powers and regions the power never controlled go invisible. It brings all noncontinuous blocks in towards the center. To do this the header bar had to be reorginzed so each box was in its own group, which ended up breaking the text centering function because I was translating the texts as a group. That code is cleaner now, as operations are run on the entire CTM instead of just the translates.