Fixed title replacement
I made it so after a hover the title of the power group goes back into its original group, not into the parent group of the box.

Added functions to change the display of an object or an array of objects to inline or none.

Changed the naming of the focus animations to focus translate animations in preparation for adding the opacity animation code