Started Configuration App + File restructuring
The start of the configuration app is in, it generates a display of the excel data, and previews of the chart and map svg (map svg insertion is commented out atm, I don't know if I want it in there. Slows everything down a lot and doesn't really give useful information). Moved all code onto my C drive and made the repo the working directory, no more copying things across